pyproject.toml
[project]
name = "genimi"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai", "vertexai", "google-cloud-aiplatform", "google",
    "python-dotenv"
]

[tool.setuptools.packages.find]
exclude = ["assets*", "benchmark*", "playground*", "scripts*", "tests*", "data*"]

[tool.wheel]
exclude = ["assets*", "benchmark*", "playground*", "scripts*", "tests*", "data*"]